refactor update question

+++ app/routes/question.js
@@ -6,15 +6,18 @@ export default Ember.Route.extend({
   },
 
   actions: {
-    destroyQuestion(question) {
-      var answer_deletions = question.get('answers').map(function(answer) {
-        return answer.destroyRecord();
-      });
-      Ember.RSVP.all(answer_deletions).then(function(){
-        return question.destroyRecord();
-      });
-      this.transitionTo('index');
+    delete(question) {
+      if(confirm('Are you sure you want to delete this question?')) {
+        var answer_deletions = question.get('answers').map(function(answer) {
+          return answer.destroyRecord();
+        });
+        Ember.RSVP.all(answer_deletions).then(function(){
+          return question.destroyRecord();
+        });
+        this.transitionTo('index');
+      }
     },
+
 
     update(question, params){
       Object.keys(params).forEach(function(key){
